Overhaul handling of generic applicaiton.
This merges the handling of a generic application `F<A0, A1, ...>` with the handling of ordinary application `f(a0, a1, ...)`.
This means that it is now possible to apply generic arguments to something that is overloaded, and we can also filter the set of overload candidates based on arity, etc.

There is a bit of kludging going on in order to try and share the logic, but this will hopefully be worth it if we ever go and add things like keyword parameters, since we can implement it once for both kinds of calls.

Note that I'd like to eventually bottlneck "subscripting" operations like `a[i0, i1, ...]` through the same machinery, but that doesn't really matter for as long as we keep those special-cased.