Don't print overload candidates when arguments are in error.
The root cause for the problem was that the logic for testing equality for `ErrorType` was wrong. This change fixes that bug, and also does some cleanup.

When we have a function-call expression where one of the arguments is an error, we don't want to get "ambiguous overload" errors that case.
Note that if we only find one reasonable candidate, we should try to check as normal (and use the result type of that candidate) to give good response.
As such, we only suppress diagnostics in the case where the call looks ambiguous.