Initial (rudimentary) support for constant folding.
This is just the front-end support for places where an "integer constant expression" needs to be supported (array sizes, generic arguments, etc.).
This is very limited, but it is enough for simple integer constant expressions to be allowed.

I also fixed up a few bugs that surfaced in test cases once the constant-folding support was added.