Refine work on "constraint solver"
This isn't a real constraint solver in its current state, but the code is trying to treat it like one.

The basic idea is that when inferring arguments for a generic, we are going to discover a bunch of constraints that will restrict the possible types for a given argument.
Rather than implement something more general, we simply store a lower and upper bound for each type parameter, where the upper bound is supposed to represent a type that the solved value can convert to (and vice versa for the lower bound).

When we have a new constraint come in, it will either affect the lower bound, upper bound, or both (in the case of an exact equality constraint).

A trait conformance constraint works a bit like an upper bound, but right now it just checks that the upper bound type satisfies the trait's requirements.

Note: this still isn't 100% correct, because we have cases like integer math ops that should *not* take floating-point input, but the default constraint solver behavior unify floating-point and integer types to floating-point types (and then fail to find the trait conformance...).
I really don't wnat to have to solve that issue, so I'm ignoring it for now.

At this point the front-end can resolve the right overload to use between generic scalar, vector, and matrix versions of builtins (rather than always applying the generic scalar one to all possible inputs).
The failure at this point is in implicit conversions not supporting scalar-to-vector despite the inference logic assuming it is supported.