Allow reflection info to be output as JSON
This is mostly being done to make it easier to create regression tests for the reflection data that we are generating.

Not all features of the reflection data are currently being output to the JSON file, but I figure I can work on those bit-by-bit once a few tests are in place.