Add a reflection test for arrays.
This showed a bug due to the way the stdlib declares functions with the same name as the builtin types, since overload resolution ends up yielding an overloaded symbol for something like `float[10]`.
I went ahead and removed those declarations from the stdlib for now, because they shouldn't be needed long-term, and they don't make sense for the rewriter anyway.