Add as reflection test for the multi-file case.
This is mostly a copy-paste of the equivalent binding-generation test.

This involved one fixup to the compiler, so that when the target is reflection JSON, we don't emit output per-translation-unit, but rather once for the whole program (collection of translation units).
The support that was added for this could be used for other formats that can support multiple translation units and/or entry points in a single container.