Fix bug with resources inside `cbuffer` declarations.
I was failing to account for the "base offset" of the cbuffer itself when emitting nested declarations, so code like this:

    cbuffer A { Texture2D x; }
    cbuffer B { Texture2D y; }

would yield output like this:

    cbuffer A : register(b0) /* : register(t0) */
    { Texture2D x : register(t0); }

    cbuffer B : register(b1) /* : register(t1) */
    { Texture2D y : register(t0);

So both `x` and `y` end up assigned register `t0`, even though the "base" texture register for `B` is `t1`. Fixing up the HLSL emit logic to accoutn for the right offsetting behavior cleared this up.

I've added a regression test for this case.