Add a disabled resources-in-structs test.
This is disabled for now because it exposed an HLSL compiler bug that bites us.
If the user declares something like:

    struct S { SamplerState a; SamplerComparisonState b; };

    S broken;

We translate it to:

    struct S { SamplerState a; SamplerComparisonState b; };

    S broken : register(s0);

And the HLSL compiler (at least the fxc I'm running) barfs on this because it tries to assign both `a` and `b` to use register `s0`, when it is *supposed* to assign them to `s0` and then `s1`.

This bug appears to bite any `struct` type that includes a combination of `SamplerState` and `SamplerComparisonState`, and there is no workaround we can do in a simple rewriter.