Add support for major HLSL buffer types in type layout and reflection.
I'm not satisfied with this approach right now because I have to change multiple places for each new buffer type (and Microsoft seems to have about a million of them).
I'd like to investigate grouping these into a bit of a hierarchy in a subsequent change (similar to what I did for `Texture_` types).

Also made a fix so that the type layout operation `GetObjectLayout()` is parameterized on a resource kind rather than the type kind.
The underlying problem is that a `StructuredBuffer` and `RWStructuredBuffer` are both buffer types, but have different resource kinds (SRV vs. UAV), so keying off the category of type isn't the right idea.
This approach will need some refinement when we start to deal with cross-compilation, because the mapping of type->resource is different on different APIs, but for now I'm leaving the code somewhat HLSL/DX-specific.