Always output whitespace between tokens when emitting raw token stream.
This is a workaround for an issue that should really be solved in the preprocessor.
When emitting raw tokens, we try to insert whitespace where it appeared in the original code,
byt tracking whether each token follows whitespace and/or a newline.

This tracking gets confused for stuff like this:

    #define foo(x) x
    return foo(y);

Because the macro invocation `foo(y)` follows whitespace, but it expands to the token `y`, and that token did *not* follow whitespace.
Thus our naive approache to emitting stuff outputs:

    returny;

which is obviously not okay.

The quick fix for now is to always output a single space before each token (unless we already output a newline), but that leads to the output being further away from what the user originally wrote.
This may get fixed more pleasantly by other changes I have in mind, though, so I don't want to dig too deep into implementing the "correct" fix.