Major overhaul of public API.
The old API had these notions:

- "compile context" that you create and then use over many compiles
- "compile result" that is the result of a single compilation (used for reflection, getting output source)
- "diagnostic sink" that is used to collect error messages

That doesn't really match the idiom for how I think clients want to interact with things, so the new approach has:

- "session" an instance of the library, that amortizes load-time operations (e.g., loading the stdlib)

- "compile request" a single compile action (similar to invoking the command-line compiler) that encapsulates the inputs to the request, the outputs (reflection/source), and collects any errors

I also axed a bunch of API that doesn't make sense in the near term, because this branch doesn't have the modularity/components features.

One other related fix here (required to get the above working cleanly) is that I replaced the representation of source files in the compilation input request to use a `SourceFile` class that can represent either an on-disk file (that the compiler/API loads for the user) or a "virtual" file that is just passed in as a string.