Changes to enable dependency info for hot reload.
Basically, we need to be able to track any and all files references during a compilation request.
This change adds that capability to the default `#include` file-find logic (it should eventually be factored out into a layer that can sit on top of a user-defined `#include` callback).

To make this work, some of the logic that was shared between the parsing/checking of the "standard library" and user code was factored out, just to simplify things (the standard library doesn't need `#incldue`, or dependency tracking, at all).
This could potentially be cleaned up more, so that everything bottlenecks through a single path.