Add rewriter support for initializer lists.
This basically just adds the parser logic, and then stubs out the semantic-checking logic for type coercion so that we never actually try to coerce stuff.
This leads to initializer lists being emitted without any checking.

A better long-term fix will probably be to add a notion of an "unparsed expression" akin to the current "unparsed statement," and use that for anything we don't need/want to deal with in the rewriter mode.