Initial support for emitting `#line` directives.
This only covers the "unparsed" tokens inside of function bodies, but that should hopefully be enough to isolate major errors and allow for step-through debugging.

A more complete change will require going through and recording complete source location information for everything the rewriter parses, so it can spit it back out as-is.

I tried to modify the corresponding test case to work with this change, but there are two issues stopping it from passing:

1. I don't seem to be getting column information 100% right (need to investigate why). At least line numbers seem okay.
2. The fxc output when I use my "pass-through" mode ends up using absolute path names, and so the output doesn't match what my line directives give. I need to implement a fix for this.