Add support for allocating UAV slots in presence of fragment outputs.
The HLSL rules for output resource registers seem to be that an `SV_Target<n>` output collides with a `register(u<n>)` binding.
I now handle this by scanning the declared outputs of any entry points and looking for outputs with the `SV_Target` semantic, and basically treating this like an explicit binding for the corresponding resource.

Note that this is done in a slightly ad hoc fashion:
- I don't look at the actual target type for the entry point, so I don't check if you are using `SV_Target` as, say, a vertex shader output
- I don't actually collect information on shader outputs into the reflection data yet (although that is a clear next step)
- I don't pretend to understand what the HLSL rules would be if you use a non-contiguous range of `SV_Target` outputs, so don't do that