Simplify `TypeLayout` representation.
The old way I was doing things split apart layout information for uniforms from the other cases, and then also tried to be "clever" by storing the layout information for up to one resource kind inline, and then attaching the rest as a linked list (to avoid allocation in the common case).

As usually happens with "clever" code, this is turning out to be harder to maintain than a simpler alternative, so I am switching to something more or less braind-dead for now, so that I can optimize later when/if it becomes important.