Add "constructors" to builtin scalar types.
Adding the declarations to the stdlib was the easy part, but actually making them work involved various fixes inside the semantic checking logic, and then the HLSL emit logic (when we actually need to deal with parsed expressions).

We are back to the point where a trivial shader can be loaded and run through the Spire front-end, including semantic analysis, and then emit to HLSL that works.