Add first end-to-end render test
The basic idea is that the `render-test` app can be run with either HLSL or Spire as the chosen language/compiler, and it will then render and capture a screenshot. We then perform an image diff between the HLSL and Spire output to make sure things are what we expect.

In order to capture screenshots and load images, we rely on `stb_image` and `stb_image_write`.