Support C-style use of `void` as parameter list
For annoying ancient legacy reasons, a bunch of C-derived languages allow a declaration like:

    int foo(void);

To declare a function `foo` with *no* parameters, instead of the more rational answer of a function with a single parameter of type `void`.

Because GLSL supports this (I think HLSL does too), and some shaders in our test corpus use it, I suppose I need to parse it.

There is a `TODO` comment hiding in the code, beacuse I do *not* want to support this arcane and backwards syntax in "pure" Spire code.