Fixes for memory leaks
The `DeclRef` type was changed at some point to include a strong reference to the declaration it identifies.
As a result, if a declaration (e.g., a struct) ever transitively included a reference to itself (e.g., a method on the struct that takes a parameter of the struct type), then a circular reference chain was created and memory would leak.

This changes puts an ordinary weak pointer in `DeclRef` to work around the issue.