Fix a bug in generic parameter substitution.
The underlying bug was in how `DeclRef` and `DeclRefType` handled substitution.
There is an optimization in place where we detect if any sub-element of a type was altered by a substitution list, and only rebuild the outer object if there was any change.
The bug was that in some cases we weren't propagating diffs in the sub-element up to the parent, so higher levels would skip substitution.

This fixes a bunch of HLSL test cases that were failing because the lack of correct substitition was making calls to the generic `mul()` operation fail (and I had yanked the non-gerneric version of `mul()` in a previous change).

I *also* went on a tear and commented or `#if 0`d out a bunch of code that is not being used right now, and also removed `const` "correctness" from a bunch of methods.
Ideally I'll come through and remove that code completely in the near future, so that some large scale renaming/refactoring steps will be easier to pull off.