Cleanup: Remove a bunch of unused code
This includes the entire pre-existing code generation strategy, along with a bunch of supporting code.
There is also some dead code in the user-facing API now that things are almost all bottlenecked through the new C API.

A future change will have to consider how much we can trim the "core" libraries down to the basics.

Yes, I know that eventually we will probably need to re-introduce an IL, but I need to get the basic front-end in shape first without as much baggage to keep updating when I change things.