GLSL: handle "arrayed" inputs for geometry shader
The inputs to a GLSL geometry shader are declared as unsized arrays.
The current workarounds I have for unsized arrays treat them as zero-size arrays for layout purposes, which ends up meaning that our logic skips allocating binding resources to them (they are zero size, right?).

I'll fix the fundamental problem sooner or later, but even with that fixed I'd still need to special-case "arrayed" inputs because the entire *array* should get a single location assigned, and not one location per element (or infinitely many locatons, for an unbounded array).