GLSL: allow specialization constant as array size
This does *not* yet deal with the case where the user does any math on the specialization constant before using it (an `IntVal` is either a literal value, or a variable reference, but nothing more complex).
The semi-obvious next step is to have `IntVal` comprise a small hierarchy of compile-time "constant" expressions (dealing with the fact that in some cases we may have a "constant" expression that isn't actually known during compilation).