Handle `struct` declaration as a type specifier
C-based languages allow for a `struct` declaration to appear stand-alone, or as a type specifier for one or more variables:

    struct Foo {};        // stand-alone
    struct Bar {} b, *bp; // as a type specifier

The existing Spire parsing logic only handled the stand-alone case, and didn't require the trailing semicolon as a result.

Some of my GLSL test cases use the type-specifier case and so I needed to update the parser to be able to handle them.
This required a few annoying changes, but overall the new way that "declarator declarations" were being handled made it reasonably easy.

Along the way I cleaned up the ad hoc handling of type name lookup in the parser, and made it instead use the full `Scope` system for lookup.
This isn't necessarily robust unless we go ahead and perform semantic checking interleaved with parsing, but it is probably good enough for now.