GLSL: Disable a test case
The test in question uses Vulkan/SPIR-V "specialization constants" in a slightly less trivial fashion than others.
The basic idiom is something like:

    layout(constant_id = 0) int FOO = 9;

    vec4 bar[FOO + 1];

The basic problem is that when asked what the size of `bar` is, we *cannot* say that there are `10` elements, because the value of `FOO` could be changed later.
That means we need to store the symbolic expression `FOO + 1` as the size of the array.

This isn't *too* bad to support in the compiler (it basically means extended the current `IntVal` hierarchy into a whole mini AST of compile-time expressions), but it adds work that doesn't really help any user in the near term.

Note: when I first started looking into this I was all worried about what this meant for the type system: e.g., is `int[FOO + 1]` the same type as `int[1 + FOO]` or even `int[FOO/2 + (FOO+3)/2]`?
Also: what should I report in the reflection meta-data for an array like that?
Luckily, the spec for Vulkan GLSL just kind of punts on these issues:

- The reflection data is apparently supposed to be based on the default value of the constant, not its symbolic value

- In terms of type equality, the only constraint is that type equality should work for an array with the size is *just* a specialization constant symbol, and not any case where arithmetic has been done. In all other cases, types using expressions on specialization constants are claimed to be non-equal (even if the expression is syntatically identical).

I don't like these rules, because they are an absolute kludge, but they at least mean that adding just enough support to stay compatible would be tractable if we decided to do it down the road.