GLSL-cross: overhaul handling of intrinsics
With this change we now associate intrinsic function definitions with an opcode that identifies the operation, and we can use these during code emission to emit an appropriate operation for the target.

This involved basically rewriting the code that generates all the builtin operation declarations.
I also made the assignment operator into a builtin (generic) declaration, just to streamline things some more.