GLSL-cross: emit `uniform` qualifier for HLSL `cbuffer`
Previously the emit logic for GLSL would just emit the qualifiers on the declaration as-is, and assume that, e.g., a constant buffer would have a `uniform` qualifier in its list.
However, this isn't true when the constant buffer came in via HLSL source code,

The new logic goes ahead and emits an appropriate qualifier a parameter block when outputting GLSL, based on the type of the parameter block itself.
In order to not emit duplicate qualifiers, we now strip out the qualifier that we used when picking the correct type during parsing (this seems reasonable since the keyword isn't really acting as a "qualifier" in these cases, and is more like a concrete piece of syntax).