Try to improve conversion ranking.
We need our conversion rankings to be good enough that we can pick the right overload for builtin math operations on arguments of different types.

The approach here isn't perfect, but I also don't want to have to implement 100% HLSL compatibility for this stuff, because implicit conversions should really be discouraged.
Also, a more robust approach to this will likely involve adding all the relevant operations to the stdlib (e.g., adding implicit conversion costs to certain "constructors"), rather than implementing them directly in the typechecker.