Fix: generate per-entry-point output even for HLSL
In the HLSL output case, we can emit a single source string that handles all the entry points requested (at least in the pass-through case), but we shouldn't neglect to also create source strings usable for each individual entry point, so that API clients who expect per-entry-point source can still request it.

TODO: I need to consider whether the notion of per-translation-unit output is all that useful at this point. I guess it is still the right mental model for the "rewriter"...