Fixup: don't double-check local declarations.
The way we are doing checking right now is a bit loose - we check declarations on-demand to help deal with out-of-order declarations.
The down-side of this is that we need to carefully mark a declaration as already having been checked, or else we might double-check it, and not all of the syntax we construct during checking is stuff that would survive a second round of checking.