Generics: don't immediately bail on unification failure.
The case here is something like:

    __generic<let N : int> float foo(vector<float,N> x);

    vector<int,3> y;
    foo(x);

What kind of error should we see, especially in the case where `foo` might be overloaded?
It seems reasonable for a user to expect that we solve that `N=3` and then call this overload by converting `y` to a `float3`.

The existing unification logic bailed on this case, because it tried to unify `int` and `float` and couldn't find a way.

There are two reasonable ways to solve this:

- The `TryUnifyTypes` logic could handle cases where one type converts to the other, and allow them to unify

- The code that *calls* `TryUnifyTypes` can ignore its failures and try to blaze ahead.

I'm still not sure which of these is right, but this change implements the latter for the case where we are specializing a generic for overload resolution.