Emit: carefully handle compound operator arguments with conversion
Given code like the following:

    float f = ...;
    int i = ...;

    i += f;

One might *think* that it compiles similar to the following:

    i = i + int(f);

but actually it is supposed to compile as:

    i = int(float(i) + f);

That is, the compiler finds the right type to use for the operation using the same rules it applies for an ordinary addition, and *then* it goes and performs converion on both sides as needed.

Our compiler handles all operators as ordinary function calls (for better or worse...), and so we see something like:

    float operator+(in out float left, float right);

    operator+(i, f);

And in this case we realize that we need to cast the `i` argument, and we generate a type-checked AST Like:

    operator+(float(i), f);

But of course a cast of a value isn't an l-value, and so if we emit code like this:

    float(i) += f;

We will generate errors in the down-stream compiler.

TODO: The right way to solve this long-term is more complex, since we really need an AST node that stores the following pieces of information:

- An l-value (`i`)
- A conversion to apply to that l-value when it is passed as the `in` argument
- A conversion to apply to the `out` argument before assigning back to the l-value

Such an expression would be a form of "complex l-value" (a non-trivial swizzle is similarly complex, for that matter). Low-level code generation would need all those pieces, while translation to a high-level language might be able to ignore them (although I don't know how well GLSL handles this stuff).