Add "subscript" declarations
The previous code around how subscript expressions (e.g., `a[i]`) were handled was a bit ad hoc, and relied on a particular notion of an "array-like type" (`ArrayLikeType`) to make it extensible. Having to encode this in the (meta)type hierarchy was a bit unfortunate, because it meant I couldn't easily add subscript support for types like textures (it also required that the index type was always just an integer).

The new approach allows types to contain subscript declarations (using a `__subscript` keyword for now), and each subcript operation defines the particular argument types it expects, along with a getter, setter or both.
For right now all the subscript operations are instrinsics, so the getters and setters have no bodies.

When a subscript operation is encountered that doesn't fit into one of our by-hand cases (TODO: get rid of all the by-hand cases) we try to look up any subscript operations in the base type, and if any are found we proceed to construct and resolve a call expression to the result of that lookup. More or less, this means an expression like:

    a[i]

is desugared into:

    a.operator[](i)

The one remaining catch is that if the subscript declaration includes a setter, then the expression is treated as an l-value (TODO: I don't handle the case where a subscript has only a setter, and no getter).

TODO: The syntax hasn't been changed to allow multiple arguments to a subscript expression, so support for multiple parameters isn't really useful right now.

To support having both setters and getters, I split the current `FunctionDeclBase` so that there is a base class `CallableDeclBase` that holds just the parameters (and result type for now), while `FunctionDeclBase` is the specialized case where there is a (single) body.
A subscript is a callable decl, but not a `FunctionDeclBase`, while the getters/setters are `FunctionDeclBase` (just without their own parameters).